hack to register subsequence as a signature
	fn:subsequence(node*, double, double) node*

besides its default signature

	fn:subsequence(item*, double, double) item*

I need this because I mix subsequence with path steps, which wont work on
item* as it requires node*, and there appears to be a bug (1280519) in
the pathfinder casting.

The more generic problem is that fn:subsequence in gerenal has the
return type as it gets in its first parameter. See the formal semantics:

http://www.w3.org/TR/xquery-semantics/#sec_fn_subsequence

Is there a way in pathfinder to do something similar to the any::1 type bindings
in MonetDB? It probably just is my botched understanding of it all.