Merged my changes from Thu, 10 Nov 2005 from the stable to the development branch:
``This commit "fixes" the symptoms of bug #1313008.  Wait for
Jan's checkin that fixes the actual cause.  He's working on it.

Reason is: The sum() function implicitly calls fn:data() on its
argument.  fn:data()'s task is to extract the content of each
item in its argument list one by one.  The current implementation
of fn:data() does not do this one by one for each item in the list,
but converts the list as a whole into a single atomic value ("23"
in your example).

You "accidentally" got the right result for your other example,
because Pathfinder tries to rewrite fn:data() into (possibly more
efficient) calls to simpler functions (typed-value() and
string-value()).  Our rewrites caught one of your cases, but not
both.  With this checkin, we will catch both of your examples,
though it does not fix the actual cause.''