guys, a bit of a hack here: - when an URL name ends in .mil pf substitutes it for .xq
this extends the already existing extension that if URIs
of imported modules end in .mil; we assume it contains the
already translated MIL output from a .xq module. Thus, we
omit translation.

this was only supported in MapiClient -lx however; this
check-in makes it 'work' also for stanhdalone pf, in that
pf goes back to the underlying .xq source.

the www.cwi.nl/~boncz/xmark/ and www.cwi.nl/~boncz/music/
query definitions have changed quite a bit using the .mil
optimization.



..for the xquery W3C standard compliancy purists:

maybe it would be better to have .xq modules in the main
URI of the import module statement, and use the 'at' as
a hint for a possible already present mil translation.

after all, the "at" is a hint according to the definition.