MonetDB KERNEL SUPPORT FOR PAGE-WISE REMAPPED MMAP-BATS =======================================================
The XML node storage scheme of MonetDB/XQuery needs some extra kernel support.

\begin{itemize}
\item
  we introduce a fast way to create BATs with the VM pages in a different order
  (to cheaply support inserting new pages halfway).
\item
  we introduce a primitive to swizzle OIDs for such remappings between RIDs (the real
  voids on disk) and PREs (the voids of the remapped copy).
\end{itemize}

Also, we use copy-on-write maps for transaction isolation. Contrary to intuition, we
write into STORE_MMAP read-only master bats (overriding the read-only protection
with a new 'force' flag), and read only from the copy-on-write copies produced by
the 'remap'. These copies provide efficient shadow paging. That is, before a master bat
is replaced into (appends are no problem) then all live read copies (copy-on-write remaps)
get a 'idempotent' replace on the affected pages (a BUN replace with a value that was
already there), such that a shadow copy of the page is made by the OS. In this way,
the subsequent update in the master bat is invisible in the copy-on-write copies
(transaction isolation). There is a 'pagesample' PROC that produces such an
idempotent

other changes were:

\begin{itemize}
\item
  gdk_bat.mx/gdk_batop.mx: a 'force' flag in update operations that allow to update
  read-only/append-only BATs anyway. Note that if used, another (WAL-based) recovery
  mechanism must be used to repair BATs after a system crash!! This is a user
  responability (in this case, of the MonetDB/XQuery frontend).
\item
  gdk_bat.mx: make a true copy of remapped BUN heaps if the bat becomes persistent or updatable.
\item
  gdk_batop.mx: optimized fetchjoin for idempotent positional joins (into a bat[void,void]
  with equal seqbase and all tuples hitting).
\end{itemize}

In swizzling as well as in positional join, specific optimizations are made for
the common case where all pages appear in order (this occurs in XML tables that have
not yet received updates or are read-only). In this case swizzling is a no-op.

For read-only tables, logical node IDs (nids) are equal to both PREs and RIDs. A
translation from NID to PRE (or RID) then becomes positional join onto a bat[void,void]
with seqbases 0@0, which is an idempotent join. This case now returns a view at
zero cost.


DETAILED CHANGES
=================

src/modules/plain/bat.mx
- rewrote remap(). it now works with mmap().
- rewrote [swizzle]() (previous name: remap_oid)
- added tests for these

src/gdk/gdk_atoms.mx
- missing cast in ATOMreplace()

src/gdk/gdk_bat.mx,v
- check for remapped heapies (hp->copied) if access mode or persistency changes
  in that case, copy them into a normal heap

src/gdk/gdk_heap.mx
- increased minimum mmap granularity to 512KB
  problem is that remap() extends the source bat with nils after the last page
  and Linux produces SIGBUS if we touch VM that does not exist (even as a hole)
  on disk. OTOH, windows NT forces us to use a minimum page (tuple) granularity
  of 64K, and given that the widest bats usined in the PRE table of MonetDB/XQuery
  are 8 bytes (64-bits systems), we need to ensure 8*64K = 512KB granularity.

  as a consequence, in MonetDB/XQuery I will increasethe threshold to start
  using mmap-bats to 256K tuples to wast always less than half (in the shredder
  but should also be done dynamically as the document grows).


src/gdk/gdk_posix.mx
- copy-on-write maps only need to open the file only read-only

src/gdk/gdk_relop.mx
- add an extra parameter 'hitalways' to batfetchjoin, that is true if the
  MonetDB app/user ensures that all oids will find a match.
  BEWARE: if this is not the case and hitalways==true is erroneously passed,
          CRASHES may occur.

  If you use the malalgebra batfetchjoin/leftfetchjoin, hitalways will be
  set to true.

  To facilitate debugging, if PROPDEBUG is set (debugmask=8, as in the
  testweb), we will not crash, rather give a warning (ie use the slower
  fetchjoin branch).

  The ultimate goal to do this is to make fetchjoins a bit faster, but
  more importantly, to be able to remove 'idempotent' fetchjoins altogether.

  Idempotent fetchjoins are joins into a bat[void,void] with equal seqbase.
  The result of such a join is equivalent to the left input *if* all tuples
  match. Such joins are heavily generated in MonetDB/XQuery in not yet modified
  XML documents when traversing the NID<->RID<->PRE mappings (which in the
  read-only case are idempotent bat[void,void] mappings).