Fix for bug #1534621 (?)
Samuel's dead code eliminator did some sanity checks on its way:
A situation like

  x := y
  ...
  y.foo()

where y.foo() has a side-effect on y usually indicates a buggy
algebra -> MIL translation.  The evaluation of any algebra
operator should not have any side-effects on its arguments (note
that the side-effect of foo() is visible in x as well).

The generated code for XMark 10 in fact lead to such a situation,
and Samuel's code validly complained.  The cause was in the
translation of the algebraic MergeUnion operator, which invoked
chk_order() on its argument if the grouping condition was only
a single column.

 -- Made the error message in mil_dce.c a bit more explanatory.

 -- Fixed the application of chk_order() in milgen.brg.

 -- Some reformatting and commenting in mil_dce.c.