Introduced the option to zap all type information from a MAL block, unless it has been marked frozen (freezeVarType()) or denotes a constant. This way the block can become subject of a new type resolution call.
This solves the hack in emptySet, which generated algebra.identity()
calls to mask the type transition from void to oid at runtime.

Code generators should be aware of this option and be judicious
in the creation of new variables.

In addition the data flow generation has been extended to deal
with potentially side-effect functions. They are recognized
as those MAL functions that return a TYPE_void.