Fix a circular dependency:
pftijah can only be used when pathfinder is loaded, and pathfinder can
nowadays only be used when pftijah is loaded (since the generated MIL
code for each and every XQuery contains pftijah code).

What we now do is require you to load pathfinder.  It is possible,
though not necessary, to load pftijah after pathfinder.  It is not
possible to load pftijah before pathfinder.  If you do this anyway,
you now get an error message that says so (in addition to a syntax
error).

This is the best we can do without completely redesigning the
internals of MonetDB4, which, obviously, we don't want to do.