Changed the schema of the database: reshred your documents!
Changed the separator for the tables that contain the combination of
URI, local name, and prefix from a colon (:) to a vertical bar (|).
The bar is not allowed in a URI or in a local name, whereas the colon
is.
Also, take the opportunity to define a single place where the value of
this separator is defined.  It is called NS_ACCEL_SEP (namespace
accelerator separator).