More verbose warnings for property checking provide more information for debugging.
Declaration/disclaimer:
-----------------------
On my FC6 desktop (64-bit, 64-bit OIDs, configured with --enable-optimize,
compiled with gcc 4.1.1), this check-in does not cause/trigger any test of
    MonetDB4, MonetDB5, template, sql (M4 & M5), pathfinder, xml, amdb
to fail that did not also fail with the most recent code base in CVS prior
to this check-in.
I did not check in detail, though, whether all previously failing tests
still fail the same way after this check-in as they did before this
check-in.