Undo -r1.195 -r1.196
This commit was intended to solve the GDK casts test on Darwin/PPC and
Solaris/Sparc.  However it does not, since the real problem is in the
PPC and Sparc (big-endian) chips representing doubles internally in such
a way that more precision is lost than on other architectures.  The
problem is hence not caused by (M5) software, but the respective chips'
double handling.