- implement delayed finalization.  Activated by the "delay-finalize" option
tijah:create-ft-index(<TijahOptions delay-finalize="100000" ...........

The number is the number of pre-nodes added to the ft-index. When less
than this number are added finalize does not modify the inverted indices.
When a user does a query on a non-finalized ft-index the index will
be finalized automatically before execution of the query.

This fix should help increase the performance of pftijah in cases where a
lot of small docs are added seperately to the index like in the streettivo
and some Wiki scripts.