Improve Pathfinder with statistical information (``statistical guide'') -----------------------------------------------------------------------
A guide is the small version of a XML-file where all equal elements in the same pathstep was merged to one element. So we needn't search through the entire

document. We can first search over guides and map that to the elements from the entire document.

The information for guide nodes will be extracted in the 'xml-shreder' and will be saved in a file "*_guide.xml". This XML-file will be passed by an option

to Pathfinder.


Example:
Document "file.xml"

<world>
    <regions>
        <region>
            <Africa/>
        </region>
        <region>
            <Asia/>
        </region>
        ...
    </regions>
</world>

will given with following command "./xml-shred -f file.xml -o out" the guide file "out_guide.xml" with following content

<node guide="1" count="1" kind="6" name="file.xml">
    <node guide="2" count="1" kind="1" name="world">
        <node guide="3" count="2" kind="3"/>
        <node guide="4" count="1" kind="1" name="regions">
            <node guide="5" count="3" kind="3"/>
            <node guide="6" count="2" kind="1" name="region">
                <node guide="7" count="4" kind="3"/>
                <node guide="8" count="1" kind="1" name="Africa"/>
                <node guide="9" count="1" kind="1" name="Asia"/>
            </node>
        </node>
    </node>
</node>

The information of guides will be stored in nasted elements with name "node" and four attributes (guide, count, kind, name)
    * guide: is a primary key
    * count: How many equal elements are at the same pathstep
    * kind: node type of the element
    * name: (optional) name of the element


The guide file can be passed by the option "pf:guide" to Pathfinder
Example:
declare option pf:guide "~/test/out_guide.xml";