Martin wins (in silence)
Make remote.exec accept/require module and function names as arguments,
instead of a string containing a function with an optional module
separated by a dot.  Do this for the sake of consistency with the rest
of MAL, and because there is no problem any more in doing so, since get
and put won't/don't deal with functions.