Instruction records are either created in isolation or are allocated on the program instruction stack directly. In the latter case, we may have to adjust their reference if we exceed the space reserved for arguments.
It does not make sense to search the complete stack,
because this would jeopardise long MAL program
build from newly allocated structures.

The alternative to this hack is to change the code in many places
and educate the programmer to not forget updating the stmtblock
after pushing the arguments. In sql_gencode this alone would be
>100 places and in the optimizers > 30. In almost all cases
the instructions have few parameters and are constructed with
the mal_builder functions.

In the end [throwing away 16hr of patching] the simple hack
with education has become the preference.