Updates to the .spec files to be ever so slightly more compatible with the Fedora packaging guidelines.
Use %configure instead of ./configure.  This implies removing the
--prefix, --bindir, etc. options.  This also implies that the compiler
will be called with -D_FORTIFY_SOURCE=2 which makes that the compiler
is even stricter than before.  So we do that now as well.

Some fixes to get things compiled with -D_FORTIFY_SOURCE=2.  Mostly,
we need to do something with the result of functions such as fwrite,
fread, ftruncate, system.

Added descriptions to all packages.

Created a MonetDB-testing package.

Enabled the MonetDB-SQL-devel package.

When creating the tar balls, remove extraneous execute bits, and don't
include files that are generated by configure (i.e. files for which
there is a corresponding .in file).