making the xmlshredder, and hence pathfinder, compile, again:
Checking for functions strdup() and strndup() in the default libraries
and providing our own version in case they are not available
is not sufficient. In case AC_CHECK_FUNCS find the functions,
we also need to make sure that the respective prototypes are declared,
e.g., be including the respective header file (string.h) in this case.
(At least) on Linux, string.h provides the prototype for strdup()
by default, but the prototype for strndup() only if _GNU_SOURCE is
defined (see `man strndup`).

Hence, in configure.ag we now first test for string.h;
then independently the availability of functions strdup() & strndup(),
and if they exist we check also the availablity of the prototype
declarations, including string.h (if available) and setting _GNU_SOURCE
(if not already set).

Likewise, in shred_helper.h, if string.h is available, we set _GNU_SOURCE
(unless it is already set), and include string.h .
Only then, we if strdup() and/or strndup() are not available, we declare our
own (local) prototype(s), and implement the function(s) in shred_helper.c .
In case strdup() and/or strndup() are available, but there prototype
declarations are not, we provide the prototypes, only.

... portability remains an never-ending story;
and no, I cannot say that I "love" configure and the autotools |-( ...