applied Sjoerd's earlier changes also to geom & datacell:
Changed the name of the generated tar balls (and the name of the
top-level directory within them).
The names are now equal to what nightly testing created, so that
doesn't have to rename the tar balls anymore.
The files are still installed in the same place as before.
Building RPMs still works.

The main change in the code is that we no longer use @PACKAGE@ since
that is the longer tar ball name which we don't want in the
installation directories we create.  Also, we can't use the defaults
when installing include files since they would also otherwise be
installed in a directory with @PACKAGE@ in it.

Note, this will undoubtedly negatively affect the monetdb-install
script, but I leave it to its owner to fix that.