[ 1928084 ] XQ: Ancestor step with debug=10 broken
broke on a left tail oid column that is first nil, then 1, then 2.
this is ordered, even though as oid: oid(nil) > 2 > 1

in reality we cosider oid always the smallerst value, and judge the
ordereness of oid columns as signed integers.

the "bug" in leftfetchjoin was that we compared oids directly in
C as oids.

maybe it would be better to use a signed type for oid after all.